@@ -41,9 +41,12 @@
     "q": "1.4.1",
     "semver": "4.3.3",
     "shelljs": "^0.8.5",
-    "sync-request": "4.1.0",
+    "sync-request": "6.0.0",
     "typed-rest-client": "^1.8.9",
-    "typescript": "2.3.4",
-    "validator": "^13.7.0"
+    "typescript": "4.0.2",
+    "validator": "^13.7.0",
+    "@types/node": "^10.17.0",
+    "@types/mocha": "^5.2.7",
+    "@types/q": "^1.5.4"
   }
 }
