@@ -34,7 +34,7 @@ process.env['ENDPOINT_DATA_ARM1_ENVIRONMENTAUTHORITYURL'] = 'dummyurl';
 process.env['ENDPOINT_DATA_ARM1_ACTIVEDIRECTORYSERVICEENDPOINTRESOURCEID'] = 'dummyResourceId';
 process.env['ENDPOINT_DATA_ARM1_SUBSCRIPTIONID'] = 'dummySubscriptionId';
 
-tr.registerMock("azure-pipelines-tasks-azure-arm-rest/azure-arm-storage", {
+tr.registerMock("azure-pipelines-tasks-azure-arm-rest-v2/azure-arm-storage", {
     StorageManagementClient: function (A, B) {
         return {
             storageAccounts: {
@@ -68,7 +68,7 @@ tr.registerMock("azure-pipelines-tasks-azure-arm-rest/azure-arm-storage", {
     }
 });
 
-tr.registerMock("azure-pipelines-tasks-azure-arm-rest/azure-arm-common", {
+tr.registerMock("azure-pipelines-tasks-azure-arm-rest-v2/azure-arm-common", {
     ApplicationTokenCredentials: function(A,B,C,D,E,F,G) {
         return {};
     }
